refactor: Separate extension tests
