+++ breeze/tests/test.ts
@@ -1,8 +1,6 @@
 import { assertEquals } from "https://deno.land/std@0.142.0/testing/asserts.ts";
 
-import { setupTwind, virtualSheet } from "../client-deps.ts";
-import breeze from "./index.ts";
-import * as extensions from "./extensions.ts";
+import breeze from "../index.ts";
 
 const onlyChildren = { children: "testing" };
 const emptySpan = { element: "span" };
@@ -16,10 +14,6 @@ const hyperlinkWithoutChildren = {
   element: "a",
   attributes: { href: "testing" },
 };
-
-const stylesheet = virtualSheet();
-
-setupTwind({ sheet: stylesheet, mode: "silent" });
 
 Deno.test("empty element", async () => {
   assertEquals(await breeze({ component: emptySpan }), "<span></span>");
@@ -404,223 +398,6 @@ Deno.test("async context evaluation for attributes", async () => {
   );
 });
 
-Deno.test("class shortcut extension", async () => {
-  assertEquals(
-    await breeze({
-      component: { element: "span", class: "demo", children: "testing" },
-      extensions: [extensions.classShortcut],
-    }),
-    '<span class="demo">testing</span>',
-  );
-});
-
-Deno.test("class shortcut visibility extension with evaluation", async () => {
-  assertEquals(
-    await breeze({
-      component: {
-        element: "span",
-        class: {
-          "font-bold": "context.href === context.pathname",
-        },
-        children: "testing",
-      },
-      extensions: [extensions.classShortcut],
-      context: {
-        href: "foo",
-        pathname: "foo",
-      },
-    }),
-    '<span class="font-bold">testing</span>',
-  );
-});
-
-Deno.test("class shortcut extension with getter", async () => {
-  assertEquals(
-    await breeze({
-      component: {
-        element: "span",
-        __class: "context.demo",
-        children: "testing",
-      },
-      extensions: [extensions.classShortcut],
-      context: { demo: "foobar" },
-    }),
-    '<span class="foobar">testing</span>',
-  );
-});
-
-Deno.test("class shortcut extension with evaluation", async () => {
-  assertEquals(
-    await breeze({
-      component: {
-        element: "span",
-        "==class": "context.demo + 'bar'",
-        children: "testing",
-      },
-      extensions: [extensions.classShortcut],
-      context: { demo: "foo" },
-    }),
-    '<span class="foobar">testing</span>',
-  );
-});
-
-Deno.test("foreach extension without context", async () => {
-  assertEquals(
-    await breeze({
-      component: {
-        element: "ul",
-        foreach: ["items", { element: "li", __children: "value" }],
-      },
-      extensions: [extensions.foreach],
-    }),
-    "<ul></ul>",
-  );
-});
-
-Deno.test("foreach extension with an array", async () => {
-  assertEquals(
-    await breeze({
-      component: {
-        element: "ul",
-        foreach: ["context.items", { element: "li", __children: "value" }],
-      },
-      extensions: [extensions.foreach],
-      context: { items: ["foo", "bar"] },
-    }),
-    "<ul><li>foo</li><li>bar</li></ul>",
-  );
-});
-
-Deno.test("foreach extension with multiple children", async () => {
-  assertEquals(
-    await breeze({
-      component: {
-        element: "ul",
-        foreach: ["context.items", [{ element: "li", __children: "value" }, {
-          element: "li",
-          __children: "value",
-        }]],
-      },
-      extensions: [extensions.foreach],
-      context: { items: ["foo", "bar"] },
-    }),
-    "<ul><li>foo</li><li>foo</li><li>bar</li><li>bar</li></ul>",
-  );
-});
-
-Deno.test("foreach extension with an array with a nested key", async () => {
-  assertEquals(
-    await breeze({
-      component: {
-        element: "ul",
-        foreach: ["context.test.items", { element: "li", __children: "value" }],
-      },
-      extensions: [extensions.foreach],
-      context: { test: { items: ["foo", "bar"] } },
-    }),
-    "<ul><li>foo</li><li>bar</li></ul>",
-  );
-});
-
-Deno.test("foreach extension with an array of objects", async () => {
-  assertEquals(
-    await breeze({
-      component: {
-        element: "ul",
-        foreach: ["context.items", { element: "li", __children: "title" }],
-      },
-      extensions: [extensions.foreach],
-      context: { items: [{ title: "foo" }, { title: "bar" }] },
-    }),
-    "<ul><li>foo</li><li>bar</li></ul>",
-  );
-});
-
-Deno.test("foreach extension with attributes", async () => {
-  assertEquals(
-    await breeze({
-      component: {
-        foreach: ["context.scripts", {
-          element: "script",
-          attributes: {
-            __type: "type",
-            __src: "src",
-          },
-        }],
-      },
-      extensions: [extensions.foreach],
-      context: {
-        scripts: [{
-          "type": "text/javascript",
-          "src": "sidewind.js",
-        }],
-      },
-    }),
-    '<script type="text/javascript" src="sidewind.js"></script>',
-  );
-});
-
-Deno.test("foreach extension with multiple scripts", async () => {
-  assertEquals(
-    await breeze({
-      component: {
-        foreach: ["context.scripts", {
-          element: "script",
-          attributes: {
-            __type: "type",
-            __src: "src",
-          },
-        }],
-      },
-      extensions: [extensions.foreach],
-      context: {
-        scripts: [{
-          "type": "text/javascript",
-          "src": "sidewind.js",
-        }, {
-          "type": "module",
-          "src": "gustwind.js",
-        }],
-      },
-    }),
-    '<script type="text/javascript" src="sidewind.js"></script><script type="module" src="gustwind.js"></script>',
-  );
-});
-
-Deno.test("foreach extension with attributes and nesting", async () => {
-  assertEquals(
-    await breeze({
-      component: {
-        "element": "html",
-        "children": [
-          {
-            "element": "body",
-            "children": [
-              {
-                foreach: ["context.scripts", {
-                  element: "script",
-                  attributes: {
-                    __type: "type",
-                    __src: "src",
-                  },
-                }],
-              },
-            ],
-          },
-        ],
-      },
-      extensions: [extensions.foreach],
-      context: {
-        scripts: [{
-          "type": "text/javascript",
-          "src": "sidewind.js",
-        }],
-      },
-    }),
-    '<html><body><script type="text/javascript" src="sidewind.js"></script></body></html>',
-  );
-});
-
 Deno.test("component lookup", async () => {
   assertEquals(
     await breeze({
@@ -793,81 +570,6 @@ Deno.test("pass render() to ==children with props in a component", async () => {
 });
 
 // TODO: Test props (Link) within composition (siteIndex -> BaseLayout)
-
-Deno.test("visibleIf causes empty render", async () => {
-  assertEquals(
-    await breeze({
-      component: { element: "span", visibleIf: "" },
-      extensions: [extensions.visibleIf],
-    }),
-    "",
-  );
-});
-
-Deno.test("visibleIf shows element based on context", async () => {
-  assertEquals(
-    await breeze({
-      component: { element: "span", visibleIf: "context.visible" },
-      extensions: [extensions.visibleIf],
-      context: { visible: true },
-    }),
-    "<span></span>",
-  );
-});
-
-Deno.test("visibleIf hides element based on context", async () => {
-  assertEquals(
-    await breeze({
-      component: { element: "span", visibleIf: "context.visible" },
-      extensions: [extensions.visibleIf],
-      context: { visible: false },
-    }),
-    "",
-  );
-});
-
-Deno.test("visibleIf shows element based on prop", async () => {
-  assertEquals(
-    await breeze({
-      component: {
-        element: "span",
-        props: { foo: true },
-        visibleIf: "foo",
-      },
-      extensions: [extensions.visibleIf],
-    }),
-    "<span></span>",
-  );
-});
-
-Deno.test("visibleIf hides element based on prop", async () => {
-  assertEquals(
-    await breeze({
-      component: {
-        element: "span",
-        props: { foo: false },
-        visibleIf: "foo",
-      },
-      extensions: [extensions.visibleIf],
-    }),
-    "",
-  );
-});
-
-Deno.test("visibleIf shows element based on context and prop", async () => {
-  assertEquals(
-    await breeze({
-      component: {
-        element: "span",
-        props: { foo: true },
-        visibleIf: "context.visible && foo",
-      },
-      extensions: [extensions.visibleIf],
-      context: { visible: true },
-    }),
-    "<span></span>",
-  );
-});
 
 Deno.test("pass utilities to ==children", async () => {
   assertEquals(
